requirements.txt
torch>=2.1.0
torchvision>=0.16.0
datasets>=2.14.0
huggingface-hub>=0.17.0
pyyaml>=6.0
wandb>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
